Bug 1774082 - Add AUMID handling for Windows toast notifications. r=nalexander
For MSIX packages, the package AUMID is used.

For NSIS/MSI installs, AUMID is retrieve from HKCR.

For portable builds, the registry is updated at runtime to populate icons, display name, and the COM server.

The COM server has been registered as will be implemented using DllSurrogate; this allows us to offload managing the eventloop and lifetime of the COM server to dllhost.exe, and minimizes the added code complexity.

The COM server is loaded as a separate executable from the sending process to prevent issues that would occur when multiple instances of the sending process (e.g. profiles running in parallel) try to stand up parallel COM servers. The separated COM executable also allows for multiple applications to share a common COM server.

Bonus: fixes Bug 1500054, Bug 1766095, and Bug 1743424.

Differential Revision: https://phabricator.services.mozilla.com/D151739